Don't downcase the emacs test directory
This is intended to test the robustness of case-insensitive file
systems. However, on case-sensitive file systems, this doesn't
work. Travis CI is using /home/travis/build/Wilfred/remacs but Emacs
would look for /home/travis/build/wilfred/remacs and fail.

This enables the xref tests to run on travis. See #99.